+++ src/backend/catalog/catalog.c
@@ -29,7 +29,7 @@ char *
 relpath(char *relname)
 {
 	char	   *path;
-	int			bufsize = 0;
+	size_t		bufsize = 0;
 
 	if (IsSharedSystemRelationName(relname))
 	{
@@ -156,7 +156,7 @@ fillatt(TupleDesc tupleDesc)
 									0, 0, 0);
 		if (!HeapTupleIsValid(tuple))
 		{
-			elog(ERROR, "fillatt: unknown atttypid %ld",
+			elog(ERROR, "fillatt: unknown atttypid %d",
 				 (*attributeP)->atttypid);
 		}
 		else
